Lab16_React_JWT/auth-django-api/requirements.txt
django==3.2
djangorestframework
djangorestframework-simplejwt
django-cors-headers